move some code from check_hooks/__init__.py to plugin-utils, move API utils to a separate top-level directory. Both will be used by telemetry API as well
