+++ api-utils/README.md
@@ -1,0 +1,5 @@
+# Api Utils
+
+Name is subject to change.
+
+FastAPI utilities, such as exceptions and [JSON:API](https://jsonapi.org/) type definitions.
