+++ check_manager/pyproject.toml
@@ -11,11 +11,12 @@ requires-python = ">=3.12"
 dependencies = [
     "cron-converter>=1.2.2",
     "eoepca-security",
-    "fastapi[standard]",
+    "fastapi[standard]>=0.115.4",
     "httpx>=0.27.2",
     "jsonschema>=4.23.0",
     "kubernetes_asyncio>=31.1.0",
-    "plugin_utils>=0.1.0",
+    "api-utils>=0.1.0",
+    "plugin-utils>=0.1.1",
     "pydantic>=2.9.2",
     "pydantic-settings>=2.8.1",
     "pyyaml>=6.0.0",
@@ -33,15 +34,16 @@ dev-dependencies = [
 ]
 
 [tool.uv.sources]
-plugin_utils = { git = "https://github.com/EOEPCA/resource-health.git", branch = "2.0.0-rc2", subdirectory = "plugin-utils" }
+api-utils = { git = "https://github.com/EOEPCA/resource-health.git", subdirectory = "api-utils" }
+plugin_utils = { git = "https://github.com/EOEPCA/resource-health.git", subdirectory = "plugin-utils" }
 eoepca-security = { git = "https://github.com/EOEPCA/python-eoepca-security.git", branch = "2.0.0-rc2" }
 
 [tool.pytest.ini_options]
 minversion = "6.0"
 addopts = "-ra -q --cov=src --cov-report=term-missing"
 testpaths = ["tests"]
-asyncio_default_fixture_loop_scope = "function"
 asyncio_mode = "auto"
+asyncio_default_fixture_loop_scope = "function"
 
 [tool.mypy]
 python_version = "3.12"
