+++ plugin-utils/pyproject.toml
@@ -1,9 +1,9 @@
 [project]
 name = "plugin-utils"
-version = "0.1.0"
+version = "0.1.1"
 description = "A hopefully growing collection of utilities to make it easier to load and use plugins for your Python application or framework."
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 
 [build-system]
 requires = ["hatchling"]
@@ -13,11 +13,14 @@ build-backend = "hatchling.build"
 dev-dependencies = [
     "mypy>=1.15.0",
     "pytest>=8.3.3",
+    "pytest-asyncio>=1.1.0",
     "pytest-cov>=6.0.0",
 ]
 
 [tool.pytest.ini_options]
 addopts = "-ra -q --import-mode=importlib --cov=plugin_utils --cov-report=term-missing"
+asyncio_mode = "auto"
+asyncio_default_fixture_loop_scope = "function"
 testpaths = ["tests"]
 
 [tool.mypy]
